[No issue][BE-김영서] fix: 좋아요 순으로 vr 메시지 조회 추가, 위도 경도 컬럼 Double로 변경

+++ msa_gradle/text/src/main/java/com/fairytail/text/vo/TextRequest.java
@@ -13,9 +13,9 @@ public class TextRequest {
 
     private String content;
 
-    private Float lat;
+    private Double lat;
 
-    private Float lng;
+    private Double lng;
 
     private Integer status;
